Assessment/Modul-16/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
python-dotenv==1.0.1
markdown==3.5.2
django-filter==24.1
